requirements.txt
fastapi
uvicorn
sqlmodel
python-jose[cryptography]
pydantic
psycopg2-binary
SQLAlchemy
